requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
pika==1.3.2
langdetect==1.0.9
databases==0.9.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
aiosqlite==0.20.0
loguru==0.7.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2


